Move StyleSheetLoader into public api folder.
